refactor: use server component and Improved performance
